# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: hollr.chat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      operationId: getMyProfile
      summary: Get my user profile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Users]
      operationId: updateMyProfile
      summary: Update my user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # Servers
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, username, displayName, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: [string, 'null']
        status:
          type: string
          enum: [online, idle, dnd, offline]
          default: offline
        customStatus:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 32
        avatarUrl:
          type: [string, 'null']
        status:
          type: string
          enum: [online, idle, dnd, offline]
        customStatus:
          type: [string, 'null']
          maxLength: 128

    # Server schemas
...